backend/requirements.txt
APScheduler==3.10.1
asgiref==3.4.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==40.0.2
DateTime==5.3
Django==3.2.20
django-apscheduler==0.6.2
django-cors-headers==3.10.1
django-dotenv==1.4.2
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
mysqlclient==2.1.1
numpy==1.26.2
openpyxl==3.1.2
packaging==21.3
pandas==1.1.5
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.0.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.1.1
tzdata==2023.3
tzlocal==4.2
urllib3==1.26.16
whitenoise==6.6.0
zipp==3.6.0
zope.interface==6.1